requirements.txt
pyjwt
yarl
pyyaml
js2py
psutil
aiohttp
nest-asyncio
cryptography